Clean resource types
