+++ tools/app/src/components/tree-nodes/ConfigTreeNode.tsx
@@ -14,7 +14,8 @@
 import { Box, Typography } from '@mui/joy';
 import { NodeRendererProps, NodeApi } from 'react-arborist';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { hasResourceData, ResourceNode } from '@/lib/api/resources';
+import { hasResourceData } from '@/lib/api/resources';
+import { ResourceNode } from '@/lib/api/types';
 import { useTranslation } from 'react-i18next';
 
 interface ConfigTreeNodeProps extends NodeRendererProps<ResourceNode> {
